+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add support for invisible tiles
 * `PartialEq` for `Tiles` now ignore internal state
 * Add `Tiles::find_pane`
+* Add `Tiles::remove_recursively`
 
 
 ## 0.1.0 - Initial Release - 2023-05-24